bug fixt

+++ src/modules/vehicles.ts/vehicles.service.ts
@@ -1,3 +1,4 @@
+import { Response } from "express";
 import { pool } from "../../config/DB";
 
 const creactVhicelDB = async (paylod: Record<string, any>) => {
@@ -7,7 +8,7 @@ const creactVhicelDB = async (paylod: Record<string, any>) => {
     return result;
 }
 const getalVhicelDB = async () => {
-    const result = await pool.query(`SELECT * FROM vehicles*`)
+    const result = await pool.query(`SELECT * FROM vehicles`)
     return result;
 }
 const getSingleVhicelDB = async (id: any) => {
@@ -19,7 +20,14 @@ const updateVhicelDB = async (paylod: Record<string, any>) => {
     const result = await pool.query(`UPDATE vehicles SET vehicle_name=$1, type=$2, registration_number=$3, daily_rent_price=$4, availability_status=$5 WHERE id=$6 RETURNING*`, [vehicle_name, type, registration_number, daily_rent_price, availability_status, paylod.params.id])
     return result;
 }
-const deleteVhicelDB = async (id: string | undefined) => {
+const deleteVhicelDB = async (id: string | undefined, res: Response) => {
+    const getalVhicel = await pool.query(`SELECT * FROM vehicles WHERE id=$1`, [id])
+    if (getalVhicel.rows.length > 0 && getalVhicel.rows[0].availability_status == 'booked') {
+        return res.status(400).json({
+            message: 'Cannot delete vehicle: It has active bookings'
+        })
+    }
+
     const result = await pool.query(`DELETE FROM vehicles WHERE id=$1 RETURNING*`, [id])
     return result;
 }
